[4837] PoC of configurable dep mgmt.
2 biggest needs.  Don't remove if has config other than GAV+exclusion
option and synchronize the version as opposed to remove.

Everything is still govered by `Comparator`.

This does not have impl for synchronize case (which is the big one we
need:)  But wanted to get something out there.  Impl is real procedural,
but biggest problem I see is the name.  Are we still going to use
'Remove..' if an option is to 'synchronize'.  I lean towards just leave
it, but I'm a fan of good names.