Annotate `Hcl.Literal#value` as `@Nullable`
This also required annotating `Hcl.VariableExpression#name` as `@Nullable`, as currently the parser has a bug and parses a `null` literal as a `Hc.VariableExpression` with a `Hcl.Identifier` for the `name` property which represents the `null` value.

Additionally, there was a similar bug in the `JsonPath` grammar which had to be fixed.