Compact hash table for string keys
Stores the bytes of strings to save some memory and in contrast to `HashMap` also doesn't store `Map.Entry` objects. The downside is slower lookups, as the hash code needs to be recomputed every time. For a given example the custom implementation saves around 15% of memory.

Uses xxHash for hashing the byte arrays.