Precondition recipes must not produce data table rows
When a recipe is used as a precondition of another recipe, it must not produce any data table rows. Only the recipes which are executed by the recipe scheduler must produce data table rows.

The original take on this (#5228) didn't work properly, because of how declarative recipes with preconditions have been implemented as classes implementing the `DelegatingRecipe` interface, where the recipe being executed by the recipe scheduler is a wrapper for the actual recipe.