New `ParenthesizeVisitor` utility (#5266)
The `Parenthesize` visitor adds any missing parentheses in a given subtree. This can be required in a variety of situations when recipes replace expressions of one type with another. So if you for example have the following Java code:

```java
boolean isString = obj instanceof String;
if (!isString)
    return;
```

then inlining the `isString` variable requires adding parentheses due to the `!` unary operator in the context where it is used.