Allow concurrent creation of `LocalMavenArtifactCache` (#5281)
* Allow concurrent creation of `LocalMavenArtifactCache`

By re-checking if the dir exists after a failure to create it allows multiple `LocalMavenArtifactCache` to be created at the same time with the same path.

If the dir exists the expression is false on the first `exists()`

If the dir does not exist and is created by the `mkdirs()` it resolves to false on that.

If the dir does not exist, then is created before `mkdirs()` can run and exists for the 2nd `exists()` it still resolves to false. Only if dir creation actually failed and it does not exist afterwards will we throw.

* Apply suggestions from code review

* Apply suggestions from code review

---------

Co-authored-by: Tim te Beek <tim@moderne.io>