Support Scanning Recipes as preconditions in DeclarativeRecipe. (#5340)
* Support Scanning Recipes as preconditions in DeclarativeRecipe.

Back in rewrite 7 we had two kinds of preconditions: Repository level and file level.
Repository level were uncommon but could be used to express things like "Run this recipe only if a particular type is in use or dependency is present".
We lost that in the jump to rewrite 8, supporting only file-level preconditions. Until now.

With these changes a DeclarativeRecipe can use a scanning recipe as a precondition. This allows for not only repository level preconditions, but for any arbitrary granularity between file and repository. So this mechanism allows you to express nuanced preconditions like "only run this recipe on java sources gradle projects which apply a particular plugin".

* A DeclarativeRecipe with no scanning preconditions need not be treated as a scanning recipe by the scheduler

* Add recipes that can be used as preconditions to only apply another recipe when a particular Gradle or Maven build plugin is present.

* Renamed RepositoryContainsFile