Continue type attribution in a case of exception (#5456)
Since type attribution is a best-effort, we continue it even in the presence of errors.
But in a current implementation for `JavaCompiler#attribute(Queue)`, it will be stopped in case of error, and not all entries in a queue will be processed.

In the fix, the attribution will be continued for the rest entries in `compiler.todo`

Fix https://github.com/openrewrite/rewrite/issues/5227