fix(maven): POM file path in deserialization exception message
Fixes #5558.

It is probably a very narrow corner case.
Reproduced with a POM file with an incorrect XML syntax:

```xml

<?xml version="1.0" encoding="UTF-8"?>
<project>
    <modelVersion>4.0.0</modelVersion>
```

Notice the empty line before <?xml opening tag.
Seems Maven parsers are very confortable with this but not Jackson Databind XML.

Here we add an argument to RawPom.parse with the Path to the pom file being parsed.
In case of exception the path is added to the exception message.

Tests adapted and 1 additional test to ensure the POM path is in the exception message.