RPC: Add support for `JavaType.Primitive`
Enums are a bit special in Java compared to many other programming languages, as they are proper objects and are allowed to have instance fields as well as implement interfaces. For normal enums the RPC mechanism only sends the enum constant's name. For `JavaType.Primitive` this doesn't quite work, because the RPC receiver currently assumes that "new" references are either transferred with a `value` or a `valueType`, but `JavaType.Primitive` instances get sent without either of these, because they have an `onChange` and a codec. Therefore, there is currently an exception for the `JavaType.Primitive` type in RPC.