Use LRU caching in RPC
Rather than having an unbounded cache or require explicit cache eviction, the `RewriteRpc` now does its own cache management. This might in turn mean that an object (or ref) could be missing and needs to be fetched via RPC.

For Java the caching is currently based on weak references and for JS we use a fixed size LRU cache and additionally also try to reduce this size whenever memory is running low.