RPC: Support `Markers` in `GetRef`
`Markers` are currently still a bit of a special case in RPC, as they have dedicated `sendMarkers()` and `receiveMarkers()` methods that need to be called. This also applies to `GetRef` as a `Markers` object is typically a "ref".

(cherry picked from commit ffa8f99619aa347fe2c7f2279e2461b881b76274)