Refactor RPC `Parse` method (#5705)
Instead of requiring the client to continue to invoke `Parse` (like is the case for `GetObject`), the `Parse` method will now return a complete list of IDs, containing one element per input that was in the request object. The client is then expected to call `GetObject` for each of these to retrieve the parsed LST model.