Sort `ResolvedPom#dependencyManagement` lazily (#5928)
The `ResolvedPom#dependencyManagement` list is sorted to allow `getManagedDependency()` to use binary search. Before this list was eagerly sorted, including in the constructor to remain backwards compatible with old data.

Now there is instead a boolean `dependencyManagementSorted` flag field to track whether the data is sorted or not. This flag also gets set to `false` when the contents are modified. Instead, the data is lazily sorted as part of the getter.