`RecipeRunStats` should cover the entire recipe run
Currently, `RecipeRunStats` uses the default configuration for the timer's distribution statistic configuration, which uses an expiry of 1 minute and a buffer length of 3. This means that the 99-percentile reported by `RecipeRunStats` will in fact only be computed by the values recorded during the last minute of the recipe run.

This commit addresses this by setting the expiry to 24 hours and also only using a buffer length of 1 (translating to one bucket in `TimeWindowPercentileHistogram`), so that there isn't any bias for recent scans/edits.