RPC: Treat changed object type as ADD (#6146)
When diffing the object graphs and serializing this to a stream of `RpcObjectData` objects, the case of an object with a changed type wasn't handled. So if, for example, a `Json.Literal` is replaced by a `Json.Identifier`, this needs special handling, as otherwise the diff will fail on the receiver side.

The solution is to send an `ADD` event rather than a `CHANGE` event.