JavaScript: Type mapping for function types
Implement type mapping for function types like in this arrow function: `(a: number, b: number): number => a + b`.

This will be mapped to a `Type.Class` instance with the fully-qualified named `𝑓` (to hopefully not collide with class definitions), where there is also a generic type arguments for the function's return type and parameters, so we get `𝑓<R, P1, P2, ...>`. The class then owns a method named `apply()` with a signature matching these generic type parameters.