Refactor formatting to do a single edit to the tree's prefix/suffixes if needed.
This allows for ChopIfTooLong and WrapIfLong to expect that parent objects will be formatted correctly during calculation of the wrapping phase.

This also allows for formatting to remove additional enters which are introduced between items that should not have an enter in between.

While this is in line with how auto-code-formatters behave (ignore if comment is present eg.), the impact downstream might be big in a lot of unit tests as they expect a certain style.