Fixing parsing of YAML tags in sequences (#6334)
* Fixes #5179

fix(yaml): prevent duplicate tags in sequences

In some YAML sequences, tags were being rendered twice: once in the
startBracketPrefix and once attached to the sequence. This caused
idempotency issues when formatting YAML.

Update YamlParser to only attach a tag if startBracketPrefix is null,
resolving the failing test case in #5179.

* Add test to verify fix

* Remove unnecessary comment

---------

Co-authored-by: Tim te Beek <tim@moderne.io>
Co-authored-by: Greg Oledzki <greg.oledzki@moderne.io>