Don't emit so very many near-duplicate rows from DependencyInsight.
I'm noticing running DependencyInsight on the OpenRewrite org (45 almost-all-gradle repositories) searching for all jackson dependencies produces a data table that is 1.7  GB on disk comprised of 18 million lines of text (counting rows was hard since the spreadsheet program crashed trying to open it).
For just rewrite-thirdparty, a single recipe module, we emit ~12,000 rows into the data table (~10MB).
If I further narrow to jackson-core alone there are still ~2,000 rows.
We seem to be exhaustively recording every single unique path to a dependency, per configuration, per project, per repository.
This is fantastically gratuitous and makes the data table so exhaustive it's often useless.