Fix error when parsing xml-s (and plain texts) encoded with UTF-8 BOM (#6420)
* Fix error when parsing xml-s (and plain texts) encoded with UTF-8 BOM
Parser.requirePrintEqualsInput() restores the BOM while printing the parsed SourceFile (in Tree.print()) but it doesn't restore the BOM on the reread input (in SourceFile.printEqualsInput()).
This causes requirePrintEqualsInput() (and XmlParser.parseInputs() which calls the equal check) to always return a ParsingError for XML-s encoded with UTF-8 BOM.
PlainTextParser suffers from the same issue, but there the parsing doesn't fail because that parser doesn't invoke the equal check during parsing.

* Duplicate the restoration of byte order mark

---------

Co-authored-by: Tim te Beek <tim@moderne.io>