Resolve existing maven properties with a null value to an empty string
Defining a maven property as `<key />` would produce a property with the
value `null`. During property replacement where it is used in a string
value this would result in an empty string.

For the placeholder helper the `getProperty` method should return an empty
string for the key, otherwise it is regarded as a non-existing property.