Resolve existing maven properties with a null value to an empty string (#6431)
* Resolve existing maven properties with a null value to an empty string

Defining a maven property as `<key />` would produce a property with the
value `null`. During property replacement where it is used in a string
value this would result in an empty string.

For the placeholder helper the `getProperty` method should return an empty
string for the key, otherwise it is regarded as a non-existing property.

* Polish test

* Use `pomXml`, not `mavenProject` to pass in XML

* Return the resolved pom if the version was updated.

* Use empty string for null property values

In versions null property values should resolve to empty strings.

* Removed dead local store

* Corrected test cases as properties in versions are now properly resolved

* Document why we're storing empty String locally

* Restore original import order

* Drop the `orElseThrow` from the assertion chain

* No need to override parent filename

* Clear out empty lines

---------

Co-authored-by: Tim te Beek <tim@moderne.io>