JavaScript: Formatter supports simple blocks/methods in one line
When formatting a simple block/method (defined as a block which is empty) the formatter now supports the same two options as the WebStorm formatter to allow controlling if these should be formatted on a single line (as `{}`) or not.

Also, the formatter now respects the `objectLiteralBraces` and `objectLiteralTypeBraces` options to control whether object and type literals should be formatted with spaces within the curly braces or not.